+++ app/src/main/java/com/nickoehler/brawlhalla/clans/presentation/ClanAction.kt
@@ -3,9 +3,9 @@ package com.nickoehler.brawlhalla.clans.presentation
 import com.nickoehler.brawlhalla.clans.presentation.model.ClanSortType
 
 sealed interface ClanAction {
-    data class SelectClan(val clanId: Int) : ClanAction
-    data class SelectMember(val memberId: Int) : ClanAction
-    data class ToggleClanFavorites(val clanId: Int, val name: String) : ClanAction
+    data class SelectClan(val clanId: Long) : ClanAction
+    data class SelectMember(val memberId: Long) : ClanAction
+    data class ToggleClanFavorites(val clanId: Long, val name: String) : ClanAction
     data class SelectSortType(val sort: ClanSortType) : ClanAction
     data object ReverseSortType : ClanAction
 }
