+++ app/src/main/java/com/nickoehler/brawlhalla/core/data/DatabaseDataSource.kt
@@ -13,15 +13,15 @@ class DatabaseDataSource(
     private val playerDao = database.playerDao()
     private val clanDao = database.clanDao()
 
-    override suspend fun savePlayer(brawlhallaId: Int, name: String) {
+    override suspend fun savePlayer(brawlhallaId: Long, name: String) {
         playerDao.insertPlayer(Player(brawlhallaId, name))
     }
 
-    override suspend fun deletePlayer(brawlhallaId: Int) {
+    override suspend fun deletePlayer(brawlhallaId: Long) {
         playerDao.deletePlayer(brawlhallaId)
     }
 
-    override fun getPlayer(brawlhallaId: Int): Flow<Player?> {
+    override fun getPlayer(brawlhallaId: Long): Flow<Player?> {
         return playerDao.getPlayer(brawlhallaId)
     }
 
@@ -29,15 +29,15 @@ class DatabaseDataSource(
         return playerDao.getAllPlayers()
     }
 
-    override suspend fun saveClan(clanId: Int, name: String) {
+    override suspend fun saveClan(clanId: Long, name: String) {
         clanDao.insertClan(Clan(clanId, name))
     }
 
-    override suspend fun deleteClan(clanId: Int) {
+    override suspend fun deleteClan(clanId: Long) {
         clanDao.deleteClan(clanId)
     }
 
-    override fun getClan(clanId: Int): Flow<Clan?> {
+    override fun getClan(clanId: Long): Flow<Clan?> {
         return clanDao.getClan(clanId)
     }
