+++ app/src/main/java/com/nickoehler/brawlhalla/core/data/database/dao/ClanDao.kt
@@ -12,11 +12,11 @@ interface ClanDao {
     fun getAllClans(): Flow<List<Clan>>
 
     @Query("SELECT * FROM clans WHERE id=:id")
-    fun getClan(id: Int): Flow<Clan?>
+    fun getClan(id: Long): Flow<Clan?>
 
     @Insert
     suspend fun insertClan(clan: Clan)
 
     @Query("DELETE FROM clans WHERE id=:id")
-    suspend fun deleteClan(id: Int)
+    suspend fun deleteClan(id: Long)
 }
