+++ app/src/main/java/com/nickoehler/brawlhalla/favorites/FavoriteAction.kt
@@ -3,9 +3,9 @@ package com.nickoehler.brawlhalla.favorites
 import com.nickoehler.brawlhalla.favorites.presentation.model.FavoriteType
 
 sealed interface FavoriteAction {
-    data class SelectPlayer(val brawlhallaId: Int) : FavoriteAction
-    data class SelectClan(val clanId: Int) : FavoriteAction
+    data class SelectPlayer(val brawlhallaId: Long) : FavoriteAction
+    data class SelectClan(val clanId: Long) : FavoriteAction
     data class SelectFavorite(val fav: FavoriteType) : FavoriteAction
-    data class DeletePlayer(val brawlhallaId: Int) : FavoriteAction
-    data class DeleteClan(val clanId: Int) : FavoriteAction
+    data class DeletePlayer(val brawlhallaId: Long) : FavoriteAction
+    data class DeleteClan(val clanId: Long) : FavoriteAction
 }
