+++ app/src/main/java/com/nickoehler/brawlhalla/legends/presentation/LegendAction.kt
@@ -4,7 +4,7 @@ import com.nickoehler.brawlhalla.legends.domain.LegendStat
 import com.nickoehler.brawlhalla.legends.presentation.models.FilterOptions
 
 sealed interface LegendAction {
-    data class SelectLegend(val legendId: Int) : LegendAction
+    data class SelectLegend(val legendId: Long) : LegendAction
     data class SelectStat(val stat: LegendStat, val value: Int) : LegendAction
     data class SelectFilter(val filter: FilterOptions) : LegendAction
     data object OnFilterToggle : LegendAction
