+++ app/src/main/java/com/nickoehler/brawlhalla/ranking/presentation/models/StatClanUi.kt
@@ -4,9 +4,9 @@ import com.nickoehler.brawlhalla.ranking.domain.StatClan
 
 data class StatClanUi(
     val clanName: String,
-    val clanId: Int,
-    val clanXp: Int,
-    val personalXp: Int,
+    val clanId: Long,
+    val clanXp: Long,
+    val personalXp: Long,
 )
 
 fun StatClan.toStatClanUi(): StatClanUi {
